+++ dkron/proc.go
@@ -80,57 +80,6 @@ func (a *AgentCommand) selectServer() serf.Member {
 	return server
 }
 
-func (a *AgentCommand) queryRPCConfig() ([]byte, error) {
-	nodeName := a.selectServer().Name
-
-	params := &serf.QueryParam{
-		FilterNodes: []string{nodeName},
-		FilterTags:  map[string]string{"dkron_server": "true"},
-		RequestAck:  true,
-	}
-
-	qr, err := a.serf.Query(QueryRPCConfig, nil, params)
-	if err != nil {
-		log.WithFields(logrus.Fields{
-			"query": QueryRPCConfig,
-			"error": err,
-		}).Fatal("proc: Error sending query")
-		return nil, err
-	}
-	defer qr.Close()
-
-	ackCh := qr.AckCh()
-	respCh := qr.ResponseCh()
-
-	var rpcAddr []byte
-	for !qr.Finished() {
-		select {
-		case ack, ok := <-ackCh:
-			if ok {
-				log.WithFields(logrus.Fields{
-					"query": QueryRPCConfig,
-					"from":  ack,
-				}).Debug("proc: Received ack")
-			}
-		case resp, ok := <-respCh:
-			if ok {
-				log.WithFields(logrus.Fields{
-					"query":   QueryRPCConfig,
-					"from":    resp.From,
-					"payload": string(resp.Payload),
-				}).Debug("proc: Received response")
-
-				rpcAddr = resp.Payload
-			}
-		}
-	}
-	log.WithFields(logrus.Fields{
-		"query": QueryRPCConfig,
-	}).Debug("proc: Done receiving acks and responses")
-
-	return rpcAddr, nil
-}
-
 // Determine the shell invocation based on OS
 func buildCmd(job *Job) (cmd *exec.Cmd) {
 	var shell, flag string
