+++ static/schema.json
@@ -180,6 +180,14 @@
           "type": [
             "object"
           ]
+       },
+        "retries": {
+          "description": "number of times to retry a failed job execution",
+          "example": 2,
+          "readOnly": false,
+          "type": [
+            "integer"
+          ]
         }
       },
       "links": [
@@ -214,6 +222,9 @@
               },
               "tags": {
                 "$ref": "#/definitions/job/definitions/tags"
+              },
+              "retries": {
+                "$ref": "#/definitions/job/definitions/retries"
               }
             },
             "required": [
